requirements.txt
streamlit
tensorflow
numpy
opencv-python-headless
Pillow
pandas